Orders: reversing sort key logic.
The way the orders contract is used now means that if an order has the 
same src amount as another, but with a smaller dst amount -> it should 
come before it.

e.g.:
Order1(src=10, dst=100)
Order2(src=10, dst=200)

Order1 means that the seller will give 10 src units when provided with 
100 dst units. This is better than Order2 as that seller is willing to 
provide 10 src units only for the higher price of 200 dst units.

So Order1 should come before Order2 in the list.