Uniswap: rate last bit: use internal inventory
Using the last bit of the rate (rate % 2) to indicate whether to use the
internal inventory or call uniswap. This way the calculation for
deciding whether to use the internal inventory happens only once in the
getConversionRate() function and the decision propogates to the trade()
function using the last bit.