Add rebate per reserve type (#900)
Add feature for Kyber team to set a flag to determine if a reserve type can get rebate

What changed? Before this commit, the rebate is given to all reserves (except the ones
whose volume don't pay any fee). From this commit on, the rebate will be given to reserves
with selected types only.

Why? Because giving volume-based rebate equally to reserves means exactly lower fee to
the takers. Rather than that, we want the rebate to encourage strategic reserves only.